Double promotion fixes.

+++ dlls/dmc/quake_items.cpp
@@ -65,7 +65,7 @@ void CQuakeItem::Spawn()
 
 	if (DROP_TO_FLOOR(ENT(pev)) == 0)
 	{
-		ALERT(at_error, "Item %s fell out of level at %f,%f,%f", STRING( pev->classname ), pev->origin.x, pev->origin.y, pev->origin.z);
+		ALERT(at_error, "Item %s fell out of level at %f,%f,%f", STRING( pev->classname ), (double)pev->origin.x, (double)pev->origin.y, (double)pev->origin.z);
 		UTIL_Remove( this );
 		return;
 	}
